+++ OgreKit/RegularExpression/OGRegularExpressionMatchPrivate.h
@@ -3,8 +3,8 @@
  * Project: OgreKit
  *
  * Creation Date: Sep 01 2003
- * Author: Isao Sonobe <sonoisa (AT) muse (DOT) ocn (DOT) ne (DOT) jp>
- * Copyright: Copyright (c) 2003 Isao Sonobe, All rights reserved.
+ * Author: Isao Sonobe <sonoisa@gmail.com>
+ * Copyright: Copyright (c) 2003-2018 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -21,9 +21,9 @@
  * 初期化 *
  *********/
 - (id)initWithRegion:(OnigRegion*)region 
-	index:(unsigned)anIndex
+	index:(NSUInteger)anIndex
 	enumerator:(OGRegularExpressionEnumerator*)enumerator
-	terminalOfLastMatch:(unsigned)terminalOfLastMatch;
+	terminalOfLastMatch:(NSUInteger)terminalOfLastMatch;
 
 - (NSObject<OGStringProtocol>*)_targetString;
 - (NSRange)_searchRange;
