+++ OgreKit/RegularExpression/OGString.h
@@ -3,8 +3,8 @@
  * Project: OgreKit
  *
  * Creation Date: Sep 22 2004
- * Author: Isao Sonobe <sonoisa (AT) muse (DOT) ocn (DOT) ne (DOT) jp>
- * Copyright: Copyright (c) 2004 Isao Sonobe, All rights reserved.
+ * Author: Isao Sonobe <sonoisa@gmail.com>
+ * Copyright: Copyright (c) 2004-2018 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -19,7 +19,7 @@ extern NSString	* const OgreStringException;
 @protocol OGStringProtocol
 - (NSString*)string;
 - (NSAttributedString*)attributedString;
-- (unsigned)length;
+- (NSUInteger)length;
 
 - (NSObject<OGStringProtocol>*)substringWithRange:(NSRange)aRange;
